Move unused rescale_coordinates to attic
